move configs for wcf client generation; use common solution (not open per genertor type);
