+++ WCF-Generator/App.config
@@ -1,17 +1,29 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
   <configSections>
+    <!-- Cofiguration of generation  wcf client-->
+    <section name="wcfClientGenerator" type="WCFGenerator.WcfClientGeneration.Configuration.WcfClientGenerator, WCF-Generator" allowLocation="false" allowDefinition="Everywhere" />
     <!-- Cofiguration of generation repositories-->
     <section name="repositoryGenerator" type="WCFGenerator.RepositoriesGeneration.Configuration.RepositoryGenerator, WCF-Generator" allowLocation="false" allowDefinition="Everywhere" />
-    <section name="serialize" type="WCFGenerator.SerializeGeneration.Configuration.SerializeConfiguration, WCF-Generator" />
+    <!-- Cofiguration of generation serialize-->
+    <section name="serialize" type="WCFGenerator.SerializeGeneration.Configuration.SerializeConfiguration, WCF-Generator" allowLocation="false" allowDefinition="Everywhere"/>
   </configSections>
   <startup useLegacyV2RuntimeActivationPolicy="true">
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.2" />
   </startup>
   <appSettings>
-    <add key="SolutionPath" value="..\..\..\..\WCF-Generator.sln" />
+    <!--<add key="SolutionPath" value="..\..\..\..\WCF-Generator.sln" />-->
+    <add key="SolutionPath" value="..\..\..\..\YumaPos.Server.WCF.sln" />
     <!--<add key="SolutionPath" value="C:\GitHub\Yumapos\WCF_3\YumaPos.Server.WCF.sln" />-->
   </appSettings>
+  <!-- Cofiguration of generation wcf client-->
+  <wcfClientGenerator>
+    <wcfServices>
+      <add ClientInterfaceName = "IBackOffice" ServiceInterfaceFileName = "IBackOfficeService.cs" />
+      <add ClientInterfaceName = "ITerminal" ServiceInterfaceFileName = "IService.cs"/>
+      <add ClientInterfaceName = "IOnlineService" ServiceInterfaceFileName = "IOnlineService.cs"/>
+    </wcfServices>
+  </wcfClientGenerator>
   <!-- Cofiguration of generation repositories-->
   <repositoryGenerator>
     <repositoryProjects>
@@ -58,6 +70,7 @@
       
     </repositoryProjects>
   </repositoryGenerator>
+  <!-- Cofiguration of generation serialize-->
   <serialize baseInterface="IStatefulObject" ignoreAttribute="SerializeIgnore" includeAttribute="SerializeInclude" generationPrefix="Do"
             mappingAttribute="MapBoToDo" mappingIgnoreAttribute="MapBoToDoIgnore">
     <projectNames>
