+++ WCF-Generator/SerializeGeneration/Helpers/SyntaxSerializationHelper.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using WCFGenerator.Common;
 using WCFGenerator.RepositoriesGeneration.Services;
 
 namespace WCFGenerator
