+++ WCF-Generator/WcfClientGeneration/Fault.cs
@@ -1,0 +1,9 @@
+namespace WCFGenerator.WcfClientGeneration
+{
+    public enum Fault
+    {
+        ValidationFault = 1,
+        UnauthorizedFault = 2,
+        NotFoundFault = 3
+    }
+}
