+++ WCF-Generator/WcfClientGeneration/ServiceDetail.cs
@@ -1,0 +1,8 @@
+namespace WCFGenerator
+{
+    public class ServiceDetail
+    {
+        public string UserName { get; set; }
+        public string FileName { get; set; }
+    }
+}
